removed the get_github_info () from step0 to the main_utils
