+++ step0_processors/step0_processor_s2_sr.py
@@ -1,9 +1,6 @@
-import os
-import json
 import ee
-import configuration as config
+from main_functions import main_utils
 from .step0_utils import write_asset_as_empty
-import requests
 
 # Pre-processing pipeline for daily Sentinel-2 L2A surface reflectance (sr) mosaics over Switzerland
 
@@ -349,7 +346,7 @@ def generate_s2_sr_mosaic_for_single_date(day_to_process: str, collection: str, 
                                                                  "MEAN_SOLAR_AZIMUTH_ANGLE"])
 
         # Getting swisstopo Processor Version
-        processor_version = get_github_info()
+        processor_version = main_utils.get_github_info()
 
         # set the extracted properties to the mosaic
         mosaic = mosaic.set('system:time_start', time_start) \
@@ -503,45 +500,3 @@ def generate_s2_sr_mosaic_for_single_date(day_to_process: str, collection: str, 
         )
         task.start()"""
 
-
-def get_github_info():
-    """
-    Retrieves GitHub repository information and generates a GitHub link based on the latest commit.
-
-    Returns:
-        A dictionary containing the GitHub link. If the request fails or no commit hash is available, the link will be None.
-    """
-    # Enter your GitHub repository information
-    owner = config.GITHUB_OWNER
-    repo = config.GITHUB_REPO
-
-    # Make a GET request to the GitHub API to retrieve information about the repository
-    response = requests.get(
-        f"https://api.github.com/repos/{owner}/{repo}/commits/main")
-
-    github_info = {}
-
-    if response.status_code == 200:
-        # Extract the commit hash from the response
-        commit_hash = response.json()["sha"]
-
-        # Generate the GitHub link
-        github_link = f"https://github.com/{owner}/{repo}/commit/{commit_hash}"
-        github_info["GithubLink"] = github_link
-
-    else:
-        github_info["GithubLink"] = None
-
-    # Make a GET request to the GitHub API to retrieve information about the repository releases
-    response = requests.get(
-        f"https://api.github.com/repos/{owner}/{repo}/releases/latest")
-
-    if response.status_code == 200:
-        # Extract the release version from the response
-        release_version = response.json()["tag_name"]
-    else:
-        release_version = "0.0.0"
-
-    github_info["ReleaseVersion"] = release_version
-
-    return github_info
